+++ chpl/chpl-resources/src/main/resources-production/jobs.xml
@@ -27,6 +27,114 @@
                 </entry>
             </job-data-map>
         </job>
+<<<<<<< HEAD
+=======
+
+        <!-- Questionable URL Data Collector -->
+        <!-- URL Status Data Collector -->
+        <!-- Job used to change status of a set listings -->
+        <job>
+            <name>updateListingStatusJob</name>
+            <group>systemJobs</group>
+            <description>Updates the status for a set of listings</description>
+            <job-class>gov.healthit.chpl.scheduler.job.UpdateListingStatusJob</job-class>
+            <durability>true</durability>
+            <recover>false</recover>
+            <job-data-map>
+                <entry>
+                    <key>parameters</key>
+                    <value>[{"label": "Status Effective Date", "key": "effectiveDate", "type": "date"}, {"label": "New Listing Status", "key": "certificationStatus", "type": "select", "items": [{"id": 1, "name": "Active"}, {"id": 2, "name": "Retired"}, {"id": 3, "name": "Withdrawn by Developer"}, {"id": 4, "name": "Withdrawn by ONC-ACB"}, {"id": 5, "name": "Pending"}, {"id": 6, "name": "Suspended by ONC-ACB"}, {"id": 7, "name": "Suspended by ONC"}, {"id": 8, "name": "Terminated by ONC"}, {"id": 9, "name": "Withdrawn by Developer Under Surveillance/Review"}] }, {"label": "Reason For Change", "key": "reason", "type": "text"}, {"label": "Email Report To", "key": "email", "type": "text"} ]</value>
+                </entry>
+            </job-data-map>
+        </job>
+
+        <!-- Job used to change status of 2014 listings to retired-->
+        <job>
+            <name>update2014ListingStatusJob</name>
+            <group>systemJobs</group>
+            <description>Updates the status for 2014 listings</description>
+            <job-class>gov.healthit.chpl.scheduler.job.Update2014ListingsStatusJob</job-class>
+            <durability>true</durability>
+            <recover>false</recover>
+            <job-data-map>
+                <entry>
+                    <key>authorities</key>
+                    <value>ROLE_ADMIN;ROLE_ONC</value>
+                </entry>
+                <entry>
+                    <key>emailCsvFileName</key>
+                    <value>2014ListingUpdateResults.csv</value>
+                </entry>
+                <entry>
+                    <key>emailSubject</key>
+                    <value>2014 Listing Update Output Results</value>
+                </entry>
+                <entry>
+                    <key>parameters</key>
+                    <value>[{"label": "Status Effective Date", "key": "effectiveDate", "type": "date"}, {"label": "New Listing Status", "key": "certificationStatus", "type": "select", "items": [{"id": 1, "name": "Active"}, {"id": 2, "name": "Retired"}, {"id": 3, "name": "Withdrawn by Developer"}, {"id": 4, "name": "Withdrawn by ONC-ACB"}, {"id": 5, "name": "Pending"}, {"id": 6, "name": "Suspended by ONC-ACB"}, {"id": 7, "name": "Suspended by ONC"}, {"id": 8, "name": "Terminated by ONC"}, {"id": 9, "name": "Withdrawn by Developer Under Surveillance/Review"}] }, {"label": "Reason For Change", "key": "reason", "type": "text"}, {"label": "Email Report To", "key": "email", "type": "text"}, {"label": "Send Update Every (x) Listings", "key": "statusInterval", "type": "text"} ]</value>
+                </entry>
+            </job-data-map>
+        </job>
+
+        <!-- Job used to change status of a single listings -->
+        <job>
+            <name>updateSingleListingStatusJob</name>
+            <group>subordinateJobs</group>
+            <description>Updates the status for a single listing</description>
+            <job-class>gov.healthit.chpl.scheduler.job.UpdateSingleListingStatusJob</job-class>
+            <durability>true</durability>
+            <recover>false</recover>
+            <job-data-map>
+                <entry>
+                    <key>authorities</key>
+                    <value>ROLE_ADMIN;ROLE_ONC</value>
+                </entry>
+            </job-data-map>
+        </job>
+        <!-- Job used to mark 2014 upload templates as deleted. -->
+        <job>
+            <name>remove2014UploadTemplates</name>
+            <group>systemJobs</group>
+            <description>Marks 2014 upload templates as deleted</description>
+            <job-class>gov.healthit.chpl.scheduler.job.Remove2014UploadTemplateJob</job-class>
+            <durability>true</durability>
+            <recover>false</recover>
+            <job-data-map>
+                <entry>
+                    <key>authorities</key>
+                    <value>ROLE_ADMIN;ROLE_ONC</value>
+                </entry>
+            </job-data-map>
+        </job>
+        
+        <!-- Questionable Activity data emailer -->
+        <job>
+            <name>Questionable Activity Report</name>
+            <group>chplJobs</group>
+            <description>Sends a report of questionable activity within the last week</description>
+            <job-class>gov.healthit.chpl.scheduler.job.QuestionableActivityEmailJob</job-class>
+            <durability>true</durability>
+            <recover>false</recover>
+            <job-data-map>
+                <entry>
+                    <key>authorities</key>
+                    <value>ROLE_ADMIN;ROLE_ONC</value>
+                </entry>
+                <entry>
+                    <key>frequency</key>
+                    <value>WEEKLY</value>
+                </entry>
+                <entry>
+                    <key>range</key>
+                    <value>7</value>
+                </entry>
+                <entry>
+                    <key>parameters</key>
+                    <value>[{"label": "Range (in days)", "key": "range", "type": "int"} ]</value>
+                </entry>
+            </job-data-map>
+        </job>
+>>>>>>> 660eb643f... update Questionable Activity Email Quartz job definition to include user params
         <!-- DELETE END MARKER -->
     </schedule>
 </job-scheduling-data>
