+++ chpl/chpl-resources/src/main/resources-staging/jobs.xml
@@ -446,8 +446,17 @@
                     <key>frequency</key>
                     <value>WEEKLY</value>
                 </entry>
+                <entry>
+                    <key>range</key>
+                    <value>7</value>
+                </entry>
+                <entry>
+                    <key>parameters</key>
+                    <value>[{"label": "Range (in days)", "key": "range", "type": "int"} ]</value>
+                </entry>
             </job-data-map>
         </job>
+        
         <trigger>
             <cron>
                 <name>qaAllBrokenSurveillanceEmailJob</name>
